tune error reporting again to prevent getting spammed by bots
ideally we should allow each system admin to choose their preferred log level in the error reporting service. But for now, most of the systems we are responsible for are too loud in raising the following errors:  
``` ruby
    'ActionController::ParameterMissing', 
     'ActionController::RoutingError',
     'ActionController::InvalidAuthenticityToken',
     'ActionDispatch::Http::MimeNegotiation::InvalidType',
     'ActiveRecord::RecordNotFound',
```
due to crawlers and bots. To ensure that we don't miss actual issues (eg: https://github.com/restarone/violet_rails/issues/578) we should limit the amount of false-positives that are reported to our bug triaging teams

Co-authored-by: donrestarone <shashikejayatunge@gmail.com>