[feature] Transcript parser plugin (#1121)
Addresses #1120 

# Transcript parser plugin

This plugin takes a video transcript and generates a clean version of it by removing numbers, timestamps, and names. Given a transcript like this:

```
1
00:00:00,300 --> 00:00:02,167
[sally]: i'm in los angeles and

2
00:00:01,920 --> 00:00:02,102
[bobby]: okay

3
00:00:03,151 --> 00:00:04,737
[sally]: maybe possibly

4
00:00:05,780 --> 00:00:13,613
[bobby78]: $ oh maybe ye so nice an is a
supporter financial supporter of hollows

5
00:00:13,345 --> 00:00:15,510
[sally]: 10 s yeah
10 hello world
hi there
$10 in my wallet
```

The plugin will parse the transcript and generate a clean version of it:

```
i'm in los angeles and okay maybe possibly $ oh maybe ye so nice an is a supporter financial supporter of hollows 10 s yeah 10 hello world hi 
there $10 in my wallet
```

![raw-and-parsed-transcripts](https://user-images.githubusercontent.com/73725297/191982630-c91e994b-d1a5-4d9a-83f0-ecc9f267eaa0.PNG)

## Metadata

The plugin requires the following metadata:

INPUT_STRING_PROPERTY: the name of the property that points to the raw unchanged transcript
OUTPUT_STRING_PROPERTY: the name of the property to which the plugin will write the parsed clean version of the transcript

- `INPUT_STRING_PROPERTY` and `OUTPUT_STRING_PROPERTY` must exist on the API resources of the API namespace that the plugin will run against
- `INPUT_STRING_PROPERTY` and `OUTPUT_STRING_PROPERTY` cannot be the same property (no overwriting is allowed)

![transcript-parser-metadata](https://user-images.githubusercontent.com/73725297/191981733-9b3b6ac5-e0f6-48e5-a73d-47786ff15c1b.PNG)

## Deciding if a transcript needs to be parsed

Each API resource should have a boolean property called `transcript_parsed`. The plugin uses this property to check if a transcript needs to be parsed. If `transcript_parsed` is `true`, then the plugin will not parse that transcript. Once the transcript of an API resource is parsed, the plugin sets its `transcript_parsed` property to `true`.

## Video demo


https://user-images.githubusercontent.com/73725297/191982004-3e257496-c2f6-481a-8ecc-1812b3bac6c6.mp4