[feature] Entity specific analytics (#1389)
* [feature] Entity specific analytics 

Addresses: https://github.com/restarone/violet_rails/issues/1292
<img width="1101" alt="Screen Shot 2023-01-04 at 8 07 53 AM" src="https://user-images.githubusercontent.com/35935196/210561884-9f2220bf-0c69-45ff-bef4-d9f9dc974e3b.png">

### Determining the visibility threshold for section views 

Threshold indicates at what percentage of the target's visibility the tracking should be executed. 0.75 threshold indicates that the tracking is done when 75% of the section is visible in the screen.

However for larger sections, 75% of the content can never be visible. Therefore we need to adjust the threshold.

A good rule of thumb is if the section is double the screen size, adjust the threshold to slightly less than 0.5 and if the section is 4 times the screen size, adjust threshold to less than  0.25, and so on

REF:

https://stackoverflow.com/questions/66296057/intersectionobserver-does-not-work-on-small-screens-for-long-sections-js/

https://developer.mozilla.org/en-US/docs/Web/API/Intersection_Observer_API#thresholds


Co-authored-by: Pralish Kayastha <50227291+Pralish@users.noreply.github.com>
Co-authored-by: Pralish Kayastha <pralishkayastha@gmail.com>